Framework updated V1.3
