+++ solutions/framework/src/main/java/utilities/helper/FailureHandling.java
@@ -1,8 +1,6 @@
-package utilities.ultils;
+package utilities.helper;
 
 import com.hansencx.solutions.logger.Log;
-import org.testng.ITestResult;
-import utilities.configuration.TestListener;
 
 /**
  * @param
