+++ solutions/regression-tests/src/test/java/com/hansencx/portal/suites/SearchTestSuite.xml
@@ -4,7 +4,7 @@
     <parameter name="browser" value="chrome" />
     <parameter name="mode" value="NonRemote" />
     <parameter name="username" value="QAREQUESTER"/>
-    <parameter name="password" value="QA!requestor1" />
+    <parameter name="encodedPassword" value="NC9CHQEUETUSBxwFXg==" />
     <listeners>
         <listener class-name="utilities.configuration.TestListener"/>
     </listeners>
