+++ solutions/regression-tests/src/test/java/com/hansencx/portal/tests/Search_Tests.java
@@ -2,14 +2,11 @@ package com.hansencx.portal.tests;
 
 import com.hansencx.solutions.portal.PortalBaseTest;
 import org.testng.Assert;
-import org.testng.ITestContext;
-import org.testng.ITestResult;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 import com.hansencx.solutions.logger.Log;
-import utilities.configuration.TestListener;
-import utilities.ultils.ExcelHelper;
-import utilities.ultils.FailureHandling;
+import utilities.helper.ExcelHelper;
+import utilities.helper.FailureHandling;
 
 /**
  * @param
